+++ lib/crossbeams/layout/renderer/select.rb
@@ -34,9 +34,9 @@ module Crossbeams
 
         def render_string(attrs, sel, disabled_option)
           <<-HTML
-          <div class="#{div_class}">#{hint_text}
+          <div #{wrapper_id} class="#{div_class}">#{hint_text}
             #{backup_empty_select}
-            <select #{attrs.join(' ')} name="#{@page_config.name}[#{@field_name}]" id="#{@page_config.name}_#{@field_name}">
+            <select #{attrs.join(' ')} #{name_attribute} #{field_id}>
             #{make_prompt}#{build_options(@field_config[:options], sel, disabled_option)}
             </select>
             <label for="#{@page_config.name}_#{@field_name}">#{@caption}#{error_state}#{hint_trigger}</label>
@@ -47,7 +47,7 @@ module Crossbeams
         def backup_empty_select
           # Hidden blank value to be submitted as a param if the Selectr box is cleared.
           return '' unless @searchable
-          %(<input name="#{@page_config.name}[#{@field_name}]" type="hidden" value="">)
+          %(<input #{name_attribute} type="hidden" value="">)
         end
 
         def make_prompt