+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- The div surrounding an input field gets a DOM id (the field's id with "_wrapper" appended). Useful to hide/show an input and its label.
 - Form can render a caption above non-remote forms.
 - Table can render a caption.
 - Table can be rendered with a top margin which can be in the range 0 to 7.
 - Link can be styled as a small button.
 ### Changed
+- Checkbox and Multi can be disabled.
 ### Fixed
 
 ## [0.1.14] - 2019-01-25